refactor(solver): drop monotone-completion publication gate; split atomic symbol registration; guard mapped-keys recursion
Scope reduction after verification: the monotone-completion gate
(complete_published_defs) pinned shared lib-def bodies at their first
finalized form. Pinned forms of mutually-referential defs can reference
each other's superseded TypeIds, turning Lazy-resolution walks into
non-converging cycles: temporal.ts (esnext.temporal+dom lib set)
stack-overflowed in extract_mapped_keys, and after guarding that walker
the same cycle re-surfaced in checker assignability evaluation
(evaluate_lazy_alias_for_assignability). The gate is not landable until
shared def bodies are canonical rather than checker-relative; the
deferred-publish experiment plumbing stays as-is. lib_augmentations/
lib_resolution completeness threading is reverted with it.

What stays (verified): atomic (symbol,file)->DefId registration, moved
to def/core/symbol_registration.rs to respect the def/core.rs shard
ceiling, and the binder lib-clone attribution fix. Witness: 6-file
NodeNext fixture, release binary, forced parallel: pre-fix ~3/50 dev
and 5/8 release runs diverged from sequential (false TS2741 on a
pre-heritage ArrayIterator form); post-fix 0/60 release byte-identical,
sequential output byte-unchanged, temporal.ts no longer crashes.

extract_mapped_keys gains a re-entrant same-TypeId guard (returns None,
the existing cannot-extract semantics): its Lazy-body recursion had no
visited set, so any mutually-referential resolution forms could
overflow the stack regardless of the gate.

AgentName: StudioOpus
