+++ crates/tsz-solver/src/def/core/symbol_registration.rs
@@ -1,0 +1,47 @@
+//! Atomic `(SymbolId, file_idx)` -> `DefId` registration for the shared
+//! [`DefinitionStore`].
+//!
+//! Per-file checkers stabilize binder symbols into `DefId`s lazily. The
+//! historical `lookup_by_symbol` -> `register` -> `register_symbol_mapping`
+//! sequence had a race window under parallel fresh checking: two checkers
+//! could both miss the lookup and each mint a distinct `DefId` for the same
+//! symbol identity. The composite index kept the last writer, but each
+//! checker cached its own `DefId` locally, splitting type identity
+//! program-wide ("Two different types with this name exist", divergent
+//! relation results — issue #13255). The entry-guarded form here makes the
+//! mint atomic: the first registrar wins and every concurrent caller
+//! converges on the winning `DefId`.
+
+use super::{DefId, DefinitionInfo, DefinitionStore};
+
+impl DefinitionStore {
+    /// Atomically resolve-or-register the `DefId` for a `(SymbolId,
+    /// file_idx)` identity. Returns the `DefId` and whether this call minted
+    /// it (`true`) or converged on an existing registration (`false`).
+    ///
+    /// `info` is only consumed when this call wins the registration; every
+    /// concurrent caller derives an identical `DefinitionInfo` skeleton from
+    /// the same binder symbol, so dropping the loser's copy is lossless.
+    pub fn register_for_symbol(
+        &self,
+        symbol_id: u32,
+        file_idx: u32,
+        info: DefinitionInfo,
+    ) -> (DefId, bool) {
+        use dashmap::mapref::entry::Entry;
+        match self.symbol_def_index.entry((symbol_id, file_idx)) {
+            Entry::Occupied(existing) => (*existing.get(), false),
+            Entry::Vacant(vacant) => {
+                // `register` touches only other maps (`definitions`,
+                // `name_to_defs`, `symbol_only_index`, `file_to_defs`),
+                // never `symbol_def_index`, so allocating under this entry
+                // guard cannot deadlock.
+                let def_id = self.register(info);
+                vacant.insert(def_id);
+                self.insert_symbol_only_mapping(symbol_id, def_id);
+                self.bump_generation();
+                (def_id, true)
+            }
+        }
+    }
+}
