+++ crates/tsz-solver/src/evaluation/evaluate_rules/mapped.rs
@@ -1446,7 +1446,33 @@ impl<'a, R: TypeResolver> TypeEvaluator<'a, R> {
     }
 
     /// Extract mapped keys from a type (for mapped type iteration).
+    ///
+    /// Re-entrant guard: the walk recurses through `resolve_lazy` bodies, and
+    /// mutually-referential def bodies (`A`'s shared-store body referencing
+    /// `Lazy(B)` while `B`'s references `Lazy(A)`) would otherwise recurse
+    /// without progress until stack overflow. Resolution forms are not
+    /// guaranteed acyclic — per-checker refinement and shared-store
+    /// publication can pin forms that point at each other — so a same-`TypeId`
+    /// re-entry returns `None` (defer), matching the existing "cannot extract
+    /// keys" semantics.
     pub(super) fn extract_mapped_keys(&mut self, type_id: TypeId) -> Option<MappedKeys> {
+        thread_local! {
+            static EXTRACT_MAPPED_KEYS_VISITING: std::cell::RefCell<FxHashSet<u32>> =
+                std::cell::RefCell::new(FxHashSet::default());
+        }
+        let entered =
+            EXTRACT_MAPPED_KEYS_VISITING.with(|visiting| visiting.borrow_mut().insert(type_id.0));
+        if !entered {
+            return None;
+        }
+        let result = self.extract_mapped_keys_impl(type_id);
+        EXTRACT_MAPPED_KEYS_VISITING.with(|visiting| {
+            visiting.borrow_mut().remove(&type_id.0);
+        });
+        result
+    }
+
+    fn extract_mapped_keys_impl(&mut self, type_id: TypeId) -> Option<MappedKeys> {
         let key = self.interner().lookup(type_id)?;
 
         let mut keys = MappedKeys {
